+++ .dev/Bootstrap.php
@@ -15,6 +15,7 @@ include_once $base . '/vendor/autoload.php';
 
 $classmap                                  = array();
 $classmap['Molajo\\Resource\\ClassMap']    = $base . '/Source/ClassMap.php';
+$classmap['Molajo\\Resource\\ClassMap']    = $base . '/Source/ClassMap.php';
 $classmap['Molajo\\Resource\\ResourceMap'] = $base . '/Source/ResourceMap.php';
 $classmap['Molajo\\Resource\\Proxy']       = $base . '/Source/Proxy.php';
 $classmap['Molajo\\Resource\\Scheme']      = $base . '/Source/Scheme.php';
