+++ .dev/Reflection.php
@@ -1,0 +1,7 @@
+<?php
+$class   = new ReflectionClass('Molajo\Resource\ClassMap');
+$methods = $class->getMethods();
+var_dump($methods);
+foreach ($methods as $method) {
+    echo '     * @covers  ' . $method->class . '::' . $method->name . PHP_EOL;
+}
