+++ .dev/Tests/ClassMapTest.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ *  ClassMap
+ *
+ * @package    Molajo
+ * @license    http://www.opensource.org/licenses/mit-license.html MIT License
+ * @copyright  2014-2015 Amy Stephen. All rights reserved.
+ */
+namespace Molajo\Resource;
+
+/**
+ * Render Test
+ *
+ * @author     Amy Stephen
+ * @license    http://www.opensource.org/licenses/mit-license.html MIT License
+ * @copyright  2014-2015 Amy Stephen. All rights reserved.
+ * @since      1.0.0
+ */
+class ClassMapTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * Setup
+     * 
+     * @return  $this
+     * @since   1.0.0
+     */
+    protected function setUp()
+    {
+        /**
+
+        $instance = new Proxy(new Resource(new MockRender));
+
+        $file  = __DIR__ . '/RenderTest.php';
+        $data  = array();
+        $stuff = $instance->render($file, $data);
+        $this->assertEquals('stuff', $stuff);
+
+         */
+
+        return $this;
+    }
+
+    /**
+     * Test Resource Renderer
+     *
+     * @return  $this
+     * @since   1.0.0
+     */
+    public function testResource()
+    {
+
+        return $this;
+    }
+}
+
+class MockClassMap
+{
+
+}
