+++ .dev/Tests/Source/A/Dog.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Test
+ *
+ * @package    Molajo
+ * @license    http://www.opensource.org/licenses/mit-license.html MIT License
+ * @copyright  2014-2015 Amy Stephen. All rights reserved.
+ */
+namespace Molajo\A;
+
+/**
+ * Test
+ *
+ * @author     Amy Stephen
+ * @license    http://www.opensource.org/licenses/mit-license.html MIT License
+ * @copyright  2014-2015 Amy Stephen. All rights reserved.
+ * @since      1.0.0
+ */
+class Dog
+{
+
+}
