+++ .dev/Tests/Source/A/Z/Stripes.txt
@@ -1,0 +1,1 @@
+Normal Text File.
