+++ .dev/Tests/Source/C/Content.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<extension type="Resource" version="1.0">
+    <name>RESOURCE_ARTICLES_NAME</name>
+    <author>Amy Stephen</author>
+    <create_date>January 2015</create_date>
+    <copyright>Copyright (C) 2014-2015 Amy Stephen. All rights reserved.</copyright>
+    <license>MIT</license>
+    <author_email>hello@molajo.org</author_email>
+    <author_url>http://molajo.org</author_url>
+    <version>1.0</version>
+    <description>RESOURCE_ARTICLES_DESCRIPTION</description>
+    <model
+            name="Articles"
+            criteria_catalog_type_id="16000"
+            criteria_extension_instance_id="16000"
+            get_customfields="1"
+            get_item_children="1"
+            use_special_joins="1"
+            check_view_level_access="1"
+            process_events="1"
+            use_pagination="1"
+            extends="Contentresources"
+            model_type="Resources"
+            data_object="Database">
+
+        <customfields>
+            <customfield name="customfields">
+                <include field="image1"/>
+                <include field="image_caption1"/>
+                <include field="image2"/>
+                <include field="image_caption2"/>
+                <include field="image3"/>
+                <include field="image_caption3"/>
+            </customfield>
+        </customfields>
+    </model>
+</extension>
