+++ .dev/Tests/Source/C/Customize.css
@@ -1,0 +1,94 @@
+/*! Customize.css v 1 | MIT License */
+
+/*! Grid row */
+.odd {
+    background-color: #DBDBDB;
+    padding-top: .5em
+}
+
+.even {
+    padding-top: .5em
+}
+
+.edit-title {
+}
+
+.edit-editor {
+}
+
+.edit-foooter {
+}
+
+.edit-sidebar1, .edit-sidebar2, .edit-sidebar3 {
+}
+
+.order-down, .order-up {
+    padding-right: 5px
+}
+
+.fi-star {
+    color: gold;
+    font-weight: bold;
+    font-size: 1.5rem;
+    text-align: center
+}
+
+.center {
+    text-align: center
+}
+
+html, body {
+    height: 100%;
+}
+
+.page-wrap {
+    min-height: 100%;
+}
+
+/*! Footer*/
+.site-footer ul.footer-links {
+    margin: 0;
+    display: inline-block;
+    padding-bottom: 0px;
+    padding-top: 10px;
+}
+
+.site-footer ul.footer-links li {
+    display: block;
+    float: left;
+    margin-left: 6px;
+    text-transform: uppercase;
+    font-weight: 700;
+    font-size: 0.75rem;
+}
+
+.site-footer ul.footer-links li a {
+    color: #666666;
+    padding: 3px 6px;
+}
+
+.site-footer ul.footer-links li a:hover {
+    color: #777777;
+}
+
+.site-footer p.copyright {
+    margin: 6px 0 0;
+    font-size: 0.75rem;
+    color: #666666;
+}
+
+.page-wrap {
+    min-height: 95%;
+}
+
+.page-wrap:after {
+    /* .push must be the same height as footer */
+    height: 80px;
+    background: #FFFFFF;
+}
+
+.site-footer {
+    /* .push must be the same height as footer */
+    height: 80px;
+    background: #333333;
+}
