+++ .scrutinizer.yml
@@ -1,0 +1,31 @@
+filter:
+    excluded_paths: [ "vendor/*", ".dev/*", "Factories/*" ]
+
+tools:
+    # https://scrutinizer-ci.com/docs/tools/external-code-coverage/
+    external_code_coverage: true
+
+    # https://scrutinizer-ci.com/docs/tools/php/php-analyzer/
+    php_analyzer: true
+
+    # https://scrutinizer-ci.com/docs/tools/php/change-tracking-analyzer/
+    php_changetracking: false
+
+    # https://scrutinizer-ci.com/docs/tools/php/code-sniffer/
+    php_code_sniffer:
+        config:
+            standard: "PSR2"
+        filter:
+            excluded_paths: [ "vendor/*", ".dev/*", "Factories/*" ]
+
+    # https://scrutinizer-ci.com/docs/tools/php/mess-detector/
+    php_mess_detector: true
+
+    # https://scrutinizer-ci.com/docs/tools/php/pdepend/
+    php_pdepend: true
+
+    # https://scrutinizer-ci.com/docs/tools/php/code-similarity-analyzer/
+    php_sim: true
+
+    # https://scrutinizer-ci.com/docs/tools/php/security-advisory-checker/
+    sensiolabs_security_checker: true
