+++ .travis.yml
@@ -4,6 +4,11 @@ php:
   - 5.4
   - 5.5
   - 5.6
+  - hhvm
+
+matrix:
+  allow_failures:
+    - php: hhvm
 
 notifications:
   irc: "irc.freenode.org#molajo"
@@ -11,7 +16,10 @@ notifications:
     - amystephen@gmail.com
 
 before_script:
-  - composer install --prefer-source
+  - composer self-update
+  - composer install
 
 script:
- - phpunit  -c .dev/phpunit.xml
+  - phpunit  -c .dev/phpunit.xml --coverage-clover=coverage.clover
+  - sh -c 'if [ "$TRAVIS_PHP_VERSION" != "hhvm" ]; then wget https://scrutinizer-ci.com/ocular.phar; fi;'
+  - sh -c 'if [ "$TRAVIS_PHP_VERSION" != "hhvm" ]; then php ocular.phar code-coverage:upload --format=php-clover coverage.clover; fi;'
