+++ Source/Adapter/AbstractAdapter.php
@@ -17,7 +17,7 @@ use CommonApi\Exception\RuntimeException;
  * @package    Molajo
  * @copyright  2014-2015 Amy Stephen. All rights reserved.
  * @license    http://www.opensource.org/licenses/mit-license.html MIT License
- * @since      1.0
+ * @since      1.0.0
  */
 class AbstractAdapter extends Cache implements ResourceInterface
 {
