+++ Source/Adapter/File.php
@@ -16,7 +16,7 @@ use CommonApi\Resource\ResourceInterface;
  * @package    Molajo
  * @copyright  2014-2015 Amy Stephen. All rights reserved.
  * @license    http://www.opensource.org/licenses/mit-license.html MIT License
- * @since      1.0
+ * @since      1.0.0
  */
 class File extends AbstractAdapter implements ResourceInterface
 {
