+++ Source/ClassMap.php
@@ -17,7 +17,7 @@ use Molajo\Resource\ClassMap\Items;
  * @package    Molajo
  * @copyright  2014-2015 Amy Stephen. All rights reserved.
  * @license    http://www.opensource.org/licenses/mit-license.html MIT License
- * @since      1.0
+ * @since      1.0.0
  */
 class ClassMap extends Items implements MapInterface
 {
@@ -50,17 +50,9 @@ class ClassMap extends Items implements MapInterface
      */
     protected function saveOutput()
     {
-        if (version_compare(PHP_VERSION, '5.4.0', '>=')) {
-            file_put_contents($this->interface_classes_filename, json_encode($this->interfaces, JSON_PRETTY_PRINT));
-            file_put_contents($this->concrete_classes_filename, json_encode($this->concretes, JSON_PRETTY_PRINT));
-            file_put_contents($this->events_filename, json_encode($this->events, JSON_PRETTY_PRINT));
-
-            return $this;
-        }
-
-        file_put_contents($this->interface_classes_filename, json_encode($this->interfaces));
-        file_put_contents($this->concrete_classes_filename, json_encode($this->concretes));
-        file_put_contents($this->events_filename, json_encode($this->events));
+        file_put_contents($this->interface_classes_filename, json_encode($this->interfaces, JSON_PRETTY_PRINT));
+        file_put_contents($this->concrete_classes_filename, json_encode($this->concretes, JSON_PRETTY_PRINT));
+        file_put_contents($this->events_filename, json_encode($this->events, JSON_PRETTY_PRINT));
 
         return $this;
     }
