+++ Source/Proxy/Uri.php
@@ -14,7 +14,7 @@ namespace Molajo\Resource\Proxy;
  * @package    Molajo
  * @copyright  2014-2015 Amy Stephen. All rights reserved.
  * @license    http://www.opensource.org/licenses/mit-license.html MIT License
- * @since      1.0
+ * @since      1.0.0
  */
 class Uri extends Scheme
 {
