+++ Source/ResourceMap.php
@@ -17,7 +17,7 @@ use Molajo\Resource\ResourceMap\Prefixes;
  * @package    Molajo
  * @copyright  2014-2015 Amy Stephen. All rights reserved.
  * @license    http://www.opensource.org/licenses/mit-license.html MIT License
- * @since      1.0
+ * @since      1.0.0
  */
 class ResourceMap extends Prefixes implements MapInterface
 {
